[MSIDisplayVersion] Wait for the child before proceeding.
When run from within the context of an MSI, Chrome's installer spins off
a child process to update the DisplayVersion field in the registry some
time after the installer completes. To mitigate a race where the child
is too slow to start, allowing msiexec to complete before it's ready,
this change adds a simple synchronization step between the parent and
child setup.exe processes. The parent now waits for the child to signal
that it's ready to go before completing the install.

Additionally, this change also propagates channel, level, and verbosity
info to the child process. This generally enables logging for better
diagnostics and ensures that any crashes and/or metrics are attributed
to the correct channel.

Bug: 1028510
Change-Id: Iff54dd638d09b56fcae034e16806838c210a5d2a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3097612
Commit-Queue: Greg Thompson <grt@chromium.org>
Auto-Submit: Greg Thompson <grt@chromium.org>
Reviewed-by: Yann Dago <ydago@chromium.org>
Cr-Commit-Position: refs/heads/main@{#913304}