Re-land: Allow multiple relayout passes when scrollbars change.
This got reverted in CL:3057505, because of a performance regression
reported as crbug.com/1232804 . To fix this, in addition to the original
CL, only allow multiple relayout passes if we're in the actual layout
pass, whereas for the measure pass, freeze both scrollbars right away
(just like without this CL). With this fix, there is no performance
regression with
perf_tests/layout/flexbox-nested-rows-and-columns-auto-overflow.html
which was added to monitor this particular issue.

Relevant parts of the original commit comment from CL:2821693 follow:

When layout of a node either causes intrinsic inline-sizes to become
dirty, or if scrollbars appear or disappear, we're allowed to relayout.
Traditionally we've frozen both scrollbars on the node and all
descendants when relaying out like this. This CL changes it so that we
only freeze the individual scrollbars that *appear* on the node, and
relayout until the scrollbar situation becomes stable. Note that this
may mean that we don't freeze any of the scrollbars at all on the node
(the node might not even have auto scrollbars, but rather got to
relayout because of intrinsic inline-size dirtiness).

Only allow freezing individual scrollbars on the node we're laying out:
keep on freezing both scrollbars on all descendants, since we cannot
make any assumptions about their logic.

This is essential in fixing crbug.com/1197029 - a second container query
match (after initial layout) would fail, because the scrollbar would be
removed and frozen in that state. See
container-queries/auto-scrollbars-001.html

Bug: 1007065, 1197029
Change-Id: I348bb83fddbb8283fd1cf74777ec78bb551ed083
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3088861
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#913343}