Speculative fix for OffscreenCanvas rendering glitch with OOPR-Canvas
This change forces a flush of the raster interface, including its
GrContext before passing rendered frames to the compositor.  This seems
like a likely fix for this bug since it shows frames with missing
rendered content.

The bug does not repro reliably on try jobs, so we'll only know for sure
whether this fix works after the CL lands by monitoring this builder:
https://ci.chromium.org/p/chromium/builders/ci/Android%20FYI%20Release%20(Pixel%204)

Bug: 1241168
Change-Id: I8324564dc572cdc9e1efcc5caeb8c81cec59ccde
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3105587
Commit-Queue: Justin Novosad <junov@chromium.org>
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#913377}