[Mac] Remove ScopedProfileKeepAlive in AppControllerMac
This ScopedProfileKeepAlive prevented the deletion of the last Profile*,
because AppControllerMac couldn't work when `_lastProfile' was null. now
that this is better supported, allow the last Profile* to be deleted
when all its browser windows are closed.

This resets the app menu to its initial state, from before the first
profile was deleted.

Bug: 1228227
Change-Id: Iacee12d5a4463d9367390bfff5298403d8b9116b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3072081
Commit-Queue: Nicolas Ouellet-Payeur <nicolaso@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/main@{#913396}