Fix Android Fullscreen Rotation with SurfaceSyncThrottle
With SurfaceSyncThrottling enabled, exiting from Fullscreen +
Picture-in-Picture mode, and changing tabs, could lead to a blank frame.
Which would occur if the previous tab had been evicted.

The Android Fullscreen + Picture-in-Picture flow differed from the
normal rotation flow, in how/when Android would notify the Browser of
screen size and orientation changes.

This change updates RenderWidgetHostViewAndroid to account for these
alternative orderings. I am going to propose a redo of the entire screen
size and orientation API, to give a centralized way for the Browser to
know of it's current state, and subsequent changes.

I've also expanded RenderWidgetHostViewAndroidRotationTest to encompass
the different event flows.

TEST=RenderWidgetHostViewAndroidRotationTest
Bug: 1223299, 1217610

Change-Id: Iabf266401be07ba0ff7565dce93a882f9a6864fc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3094826
Reviewed-by: Bo <boliu@chromium.org>
Commit-Queue: Jonathan Ross <jonross@chromium.org>
Cr-Commit-Position: refs/heads/main@{#913404}