ChromeVox: Ensure the panel stays in sync with ChromeVox enabled state
This change fixes an issue where after signing in during oobe, the
panel got out of sync with ChromeVox and disappeared. During the
profile switch after sign-in, we were getting into a state where
the panel was created when ChromeVox was off. It was later destroyed
when ChromeVox was turned back on.

To fix this issue, we ensure that ChromeVox is on and there is no
active panel when creating a new one.

after sign-in during OOBE.

AX-Relnotes: Fixed an issue where the ChromeVox panel disappeared
Fixed: 1241042
Change-Id: I34f14ab724e8208ae7fa48d58aae42e78930e9c7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3103596
Commit-Queue: Akihiro Ota <akihiroota@chromium.org>
Reviewed-by: David Tseng <dtseng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#913456}