media: Update SetOutputParams() (mojom/C++) to take a callback
The purpose of the callback is for MediaFoundationRenderer to notify
MediaFoundationRendererClient to use the new DCOMP texture size after
the MediaFoundationRenderer calls IMFMediaEngineEx::UpdateVideoStream().
This ensures a new size is updated on the DCOMPTexture (in GPU process)
so that SwapChainPresenter will add the video to the DirectComposition
visual tree. This is to avoid adding the video to the DirectComposition
visual tree with a wrong size at the beginning of the playback.


This CL also ensures:
- output_size is for updating DCOMPTexture size
- natural_size is the size for creating VideoFrame

Bug: 999747
Change-Id: If024f553a5589c32419a02b679b4387e4c92f498
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3091365
Reviewed-by: Sunny Sachanandani <sunnyps@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Xiaohan Wang <xhwang@chromium.org>
Commit-Queue: Frank Li <frankli@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#913479}