Update WebrtcVideoEncoderGpu to use NV12 instead of I420
Back when this class was written the VideoEncodeAccelerator only
accepted frames that used I420. Since then, the VEA was updated to
prefer NV12 so if you provided a frame which used I420, it would
convert it to NV12.

In our encoder, we converted ARGB to I420 and then the VEA converted
from I420 to NV12. This CL updates our encoder to convert from ARGB to
NV12 so we can skip the extra conversion.

Bug: b:176820162
Change-Id: I7d6d8fd28324a38221eabcb419949d582d6c3d29
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3105163
Auto-Submit: Joe Downing <joedow@chromium.org>
Commit-Queue: Lambros Lambrou <lambroslambrou@chromium.org>
Reviewed-by: Lambros Lambrou <lambroslambrou@chromium.org>
Cr-Commit-Position: refs/heads/main@{#913492}