Omnibox Security Indicators: Disable promo when lock icon policy is set
The in-product help for the new connection security indicators is meant
to be suppressed if the old indicator (lock icon) is still being shown.
A new enterprise policy that forces the lock icon to be used was added
in https://crrev.com/c/2893573 but the IPH suppression was still only
consulting the Feature (kUpdatedConnectionSecurityIndicators). This CL
updates the IPH suppression logic to consult both.

Bug: 1241535
Change-Id: Iedc2e06902307261a1c1dabbae71f233c5550eba
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3103855
Commit-Queue: Mustafa Emre Acer <meacer@chromium.org>
Reviewed-by: Justin Donnelly <jdonnelly@chromium.org>
Cr-Commit-Position: refs/heads/main@{#913511}