[LayoutNG] Unsplittable scrolling overflow in LayoutNG
According to the spec "UAs may consider as monolithic any elements
with overflow set to auto or scroll and any elements with overflow:
hidden and a non-auto logical height (and no specified maximum
logical height)." https://www.w3.org/TR/css-break-3/#possible-breaks

We previously did not treat any element with 'overflow: hidden' as
monolithic. We also would allow elements with an overflow of auto
or scroll to break if the block had auto logical height, auto or
undefined max-logical-height and a zero or auto min-logical-height.

This CL updates this logic to treat any element with overflow auto,
scroll, or hidden as monolithic (unless we're printing). And this
logic is updated specifically for LayoutNG. In legacy layout, this
would not have worked because even if an element is treated as
monolithic, it can still get chopped off if it did not fit within
the fragmentainer.

The new behavior is compatible with Firefox. The existing non-WPT
tests that are no longer relevant as a result of the updated
behavior have been deleted:
fast/multicol/composited-relpos-clipped.html
fast/multicol/composited-relpos-in-clipped.html
fast/multicol/overflow-across-columns.html

One WPT test was updated to use 'overflow: clip' instead of
'overflow: hidden' to maintain the test's current behavior:
external/wpt/css/css-multicol/multicol-width-small-001.xht

Three WPT tests were added to test the new functionality, and two
new crashing WPT tests were added that are fixed by this CL.

Bug: 1240384, 829028
Change-Id: I64b891662a2004c8da2cc20d6a0169d1a58b64c8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3103651
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Commit-Queue: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#913518}