Reland "[Start] Hide both Grid and carousel Tab switcher after leaving Start"
The original CL got reverted because the call of
TabSwitcherMediator#postHiding() is moved to the
TasksSurfaceCoordinator, while TasksSurface is null when the Start
surface is disabled. In this CL, we add a check of whether the Start
surface is enabled or not, and if not, still calls
getGridTabListDelegate().postHiding() in the StartSurfaceLayout.

The original CL description:
In this CL, we unifies the soft cleaning of Tab switchers. After
leaving the Start surface, both Grid and carousel tab switchers will
be cleaned up if created.

Bug: 1239669, 1241426
Change-Id: I456fdfab7fc67ad5ad866a76c91d31c3e4942ef1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3107510
Commit-Queue: Xi Han <hanxi@chromium.org>
Reviewed-by: Wei-Yin Chen (陳威尹) <wychen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#913549}