Side panel: Start bookmark drag and drop implementation
BookmarkFolderElement is responsible for adding draggable attributes to
the draggable elements within itself and exposes a function for
retrieving data related to the bookmark being dragged.

BookmarksDragManager listens for drag events on the delegate, the
BookmarksListElement, and creates a DragSession from the element being
dragged. DragSession is then responsible for making the appropriate
API calls and updating the UI as the drag events continue.

This CL only listens for dragstart events and calls the
chrome.bookmarkManagerPrivate.startDrag method to initiate the drag,
and does not yet include any dragging or dropping within the side panel.
Initiating the drag only allows users to start the drag from the side
panel and finish the drag process in the Bookmarks Bar or the Bookmarks
Manager. Future CLs will fill in the missing implementation for
allowing users to continue the drag process within the side panel.

Bug: 1234182
Change-Id: I8fb9311f63fa87773da194a636eb2f86aa80354a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3105105
Reviewed-by: Rebekah Potter <rbpotter@chromium.org>
Commit-Queue: John Lee <johntlee@chromium.org>
Cr-Commit-Position: refs/heads/main@{#913566}