Fixed the bounce up effect of the shelf after swipe down gesture
In tablet mode, the shelf leaves hidden_shelf_in_screen_portion pixels
wide in the screen for gesture when it is hidden in the AUTO_HIDE state.
Swiping the shelf to hide will aim such bounds as the target bounds.
However, if the end drag position is lower than the hidden shelf bounds,
the shelf will bounce upward and disappear with the opacity set to 0.
Solved this bug by directly set the shelf opacity to 0 and move the
shelf to the target bounds without animation if that happens. Note that
the hiding transition should be smooth enough as there are only a few
pixels high that will vanish immediately without animation.

Bug: 1203861
Change-Id: Ib700f5280a6d66a07cfb6a4852152e0b5f02787c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3077131
Reviewed-by: Toni Baržić <tbarzic@chromium.org>
Commit-Queue: Wen-Chien Wang <wcwang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#913574}