ScriptRunner: Simplify async script execution.
This CL is a follow-up to https://crrev.com/c/3090925 , and unifies the
async PendingScript execution to `ScriptRunner::ExecutePendingScript()`.

Before this CL, scheduling an async script execution used an mechanism
that predates closure. Previous logic involved queuing the ready
PendingScripts to `async_scripts_to_execute_soon_` deque and
`ScriptRunner::ExecuteTask()` picking up the script from the deque.
This CL utilizes `WTF::Bind` to create a closure that contains a strong
reference to the ready async PendingScript and enqueue that to the
`task_runner_`.

Bug: 1239245
Change-Id: I845ae9e1b5e54a2b56528618cfd04ac634c34a42
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3099550
Reviewed-by: Hiroshige Hayashizaki <hiroshige@chromium.org>
Commit-Queue: Kouhei Ueno <kouhei@chromium.org>
Cr-Commit-Position: refs/heads/main@{#913626}