Prevent OOB on dragging tab group by checking Contents Data size
On windows in touch mode, when dragging a tab group header, it is
possible for a web page to close itself. If that happens, the contents
data size will change and the tab strip will start animating out its
tab, however in the case of TabGroups when dragging the group header,
first it kicks the tab out of the group, changing the index, then it
closes the tab. If the header is dragged past that closing group and
the user ends the drag session, then index of the tab group header may
be incorrect, causing a UaF. To prevent the security issue, a check
is added before accessing contents data that blocks invalid index
access.

Bug: 1239057
Change-Id: I026c11ba4b67219d6c823ed286a09ee6bf564f5c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3108409
Commit-Queue: David Pennington <dpenning@chromium.org>
Reviewed-by: Peter Boström <pbos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#913631}