[Categorical search] Change handling of counts in score normalizer
The score normalizer saves some state as counts in bins, and the score
values that delineate those bins. This looks like the following:

    counts:   5 9 7 4 2 1 7
    dividers:  | | | | | |

Currently the counts are initialized to a vector of zeros of the max size,
and the dividers are initialized empty and added as-needed. This causes
a state inconsistency when writing and then reading from prefs which
fails a DCHECK.

I've changed the reservoir to initialize counts to [0.0], and grow
the counts alongside the dividers. I've added a lot of DCHECK calls in
to ensure the invariant |counts| = |dividers|+1 always holds.  I've
also made a couple of other changes while here:

1. Change the default-returned score from 1 to 0.5 if there's no data
   in the normalizer, as the 'least informative' score.

2. If the size of dividers read from disk doesn't match the max number
   set by the ctor, then we've probably updated the code to use a
   different number of bins. In this case just clear the state.

3. Fixed some lint errors.

Bug: 1199206
Change-Id: I47d352518097b37415f871f7e56a3dabbda1697b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3086800
Commit-Queue: Tony Yeoman <tby@chromium.org>
Reviewed-by: Rachel Wong <wrong@chromium.org>
Cr-Commit-Position: refs/heads/main@{#913668}