Ash shelf: Make implicit dependencies of PinAppWithID explicit.
This CL is a refactor with no intended behavior change.

The ash shelf has a method PinAppWithID which will either pin an
existing item, or create a new item, pin it, and implicitly create a
delegate for the item. This CL makes this last step of delegate creation
explicit.

This CL creates a new class ChromeShelfItemCreator that is responsible
for creating a ShelfItem and ShelfItemDelegate from an app_id. All
callsites in //chrome are rerouted to use this new class. We add an
accessor to this class from ChromeShelfController to avoid creating yet
another global, as the callsites are already using the
ChromeShelfController global.

This is one callsite in //ash that implicitly relies on asynchronous
ShelfItemDelegate creation -- that is when an app is dragged from the
app_list to the shelf. Unfortunately, the existing abstractions mean
that the app_list code does not know anything about the
app-being-dragged other than its app_id. We explicitly route a
ShelfItemCreator delegate through ShelfModel for this one use case. We
mark the method as deprecated to avoid future additional users.

There is currently an implicit dependency of ChromeShelfItemCreator on
ChromeShelfController via the constructor of
AppShortcutShelfItemController, which uses the ChromeShelfController
global to get the active profile. This can change due to multi-login, a
feature in maintenance mode. This is not a hard dependency and should
eventually be severed by refactoring AppShortcutShelfItemController to
not depend on ChromeShelfController, but this is not a problem we need
to solve now.

Change-Id: I4ff2cb7b53f06ad8e4be14f2711eb663ffcb1972
Bug: 1238470
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3088193
Commit-Queue: Erik Chen <erikchen@chromium.org>
Reviewed-by: James Cook <jamescook@chromium.org>
Reviewed-by: Toni Baržić <tbarzic@chromium.org>
Reviewed-by: Peter Boström <pbos@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Cr-Commit-Position: refs/heads/main@{#913669}