[RBD] Avoid appending multiple utm_source tags
The utm_source tag is appended when the Cart module is created to cover
the no discount case. When there is discount for the Cart, utm_source is
appended after the encoded token url. In some cases, there are two
utm_source tags appended. This CL updates the #AppendUTM to use
net::AppendOrReplaceQueryParameter instead of net::AppendQueryParameter,
so we can assure only one utm_source is used.

Change-Id: I9fd985552eac639860f4aa96a0fe8a56018c1f6d
Bug: 1241513
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3108057
Commit-Queue: Mei Liang <meiliang@chromium.org>
Reviewed-by: Wei-Yin Chen (陳威尹) <wychen@chromium.org>
Reviewed-by: Yue Zhang <yuezhanggg@chromium.org>
Cr-Commit-Position: refs/heads/main@{#913674}