Remove CRWJSInjector
This CL removes CRWJSInjector, which is part of an old JavaScript
feature architecture and no longer used extensively. It also relies on
the deprecated CRWJSInjectionEvaluator.

It keeps all the application logic the same, but moves the injection of
the windowID and execution of JavaScript into CRWWebController. The web
controller already owns the WKWebView, so it can implement the
JavaScript APIs without passing the web view to other classes.

Bug: 905939
Change-Id: I73be5b007e9aef9ece564bd0f7d2a64f78423f94
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3102865
Commit-Queue: Mike Dougherty <michaeldo@chromium.org>
Reviewed-by: Ali Juma <ajuma@chromium.org>
Cr-Commit-Position: refs/heads/main@{#913686}