FastPair: Use binary proto format for metadata.
Now that the binary proto format is available from the server
this change switches from requesting JSON to requesting and
parsing the response as a binary proto instead.  This reduces
the size of the payload and simplifies the code required to
parse the response.

Fixed: 1234546
Change-Id: I766e48aa13eb95f0737daf0a6915a861a53d8946
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3100554
Auto-Submit: Jon Mann <jonmann@chromium.org>
Commit-Queue: Jon Mann <jonmann@chromium.org>
Reviewed-by: Shane Fitzpatrick <shanefitz@google.com>
Cr-Commit-Position: refs/heads/main@{#913689}