Remove AutomationListener on page navigation
This fixes the issue where the extension doesn't get correct AXTreeData
after the extension page reloads. This resolves the flaky test result of
Tast accessibility tests.

Before this CL, AutomationEventRouter only observes RenderProcessHost.
However, when the extension's background page reloads, the same
RenderProcessHost is reused and AutomationEventRouter continues
dispatching the event to the render. On the renderer side,
AutomationMessageFilter in AutomationInternalCustomBindings listens to
the event, but it's lazily initialized on the first getDesktop or
getTree invocation. Thus the event that the browser process sent between
the reload and getDesktop are dropped. Because automation only
serializes the differences of trees, some tree data is not correct in
the extension render side.

With this CL, AutomationEventRouter removes the internal listeners list
on the navigation of the extension page. By doing this, the next time
the listener is registered,
AutomationEventRouterObserver::ExtensionListenerAdded is invoked and the
next event serializes the entire tree data.

Bug: 195363367
Test: Tast arc.Accessibility* a11y.*
Change-Id: Ia708735f81f788fbfad491d9cdf6f35d9835442b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3104664
Commit-Queue: Hiroki Sato <hirokisato@chromium.org>
Reviewed-by: David Tseng <dtseng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#913697}