[block-in-inline] Don't force non-empty for synthesized forced break
List markers can force the line to be non-empty if it is the
last line. However, a synthesized forced break by block-in-
inline is considered as if it is wrapped with an anonymous
block:
```
  <span>
    <div></div>
  </span>
```
is similar to:
```
  <anon-block><span></span></anon-block>
  <anon-block><div></div></anon-block>
  <anon-block><span></span></anon-block>
```
and that the list markers should not force the first `<anon-
block>` non-empty.

When NGBlockInInline is enabled, fixes:
  marker-before-empty-inline.html
and some more.

Bug: 716930
Change-Id: I73700090fbb78eb0075d317c6c5d947ff30865fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3106630
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Commit-Queue: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#913698}