Fix certificate upload crash on unregistered client.
EnrollmentCertificateUploaderImpl checked if CloudPolicyClient is
registered only on first attempt but not on retry. The client might
become unregistered while uploader in retry cycle.

Move is_registered check so the uploader checks the client not only on
start but also on retry. Add unit tests to cover this flow.

Bug: chromium:1241177
Test: unit_tests --gtest_filter="EnrollmentCertificateUploaderTest*"
Change-Id: I2fcb5964dbb68972a48ce72ee323bd29378acb4d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3107106
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Commit-Queue: Artem Sumaneev <asumaneev@google.com>
Cr-Commit-Position: refs/heads/main@{#913753}