Allow clearing jxl animation frames
Remove the override that disabled ClearFrameBuffer and make use of
the rewinding and skipping feature of the JXL decoder to decode an
older animation frame from the beginning efficiently.

Before this change, the image decoder was remembering all frames,
even if they were requested to be cleared, which cost a lot of
memory. Decoding from the beginning is now efficient enough to not
need to do that.

Bug: 1211339
Change-Id: I9d1f9038e23ea60dc7c24af0f6d049417b13f4b5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2953140
Commit-Queue: Lode Vandevenne <lode@google.com>
Reviewed-by: Leon Scroggins <scroggo@google.com>
Cr-Commit-Position: refs/heads/main@{#913759}