[attribution_reporting] Use single callback for ConversionReporter
Instead of one per conversion ID.

Prior to this CL, if two reports for the same conversion ID were added
to the ConversionReporter, the earlier one's callback would be dropped
and never run. For timed sends, this is fine, as the only action
performed by the callback is deleting the report, for which it doesn't
matter which of the two callbacks is run, since their behavior is the
same. But if at least one of the callbacks corresponds to a
send-all-reports-now action triggered by the internals UI, then the
callback acts as a barrier that ensures that the UI is only refreshed
once all the reports have been sent. If that callback isn't run, then
the UI will never automatically refresh.

This CL makes this situation impossible, by using only a single callback
for the ConversionReporter, specified at construction time, and removing
the internals UI barrier. Instead, the ConversionManager tracks the
outstanding report IDs; once the count reaches zero, it runs the
internals UI callback.

To simplify the implementation further, we allow only a single
outstanding send-all-reports-now action, as we expect multiple
simultaneous calls to be rare and without benefit.

Change-Id: I2a555d78a5740da03bdf64cd1ec87d7756e93dd6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3107568
Commit-Queue: Andrew Paseltiner <apaseltiner@chromium.org>
Reviewed-by: John Delaney <johnidel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#913775}